Preparing company data tests.
