+++ testing/Integrity/FromCsvTestCase.php
@@ -1,0 +1,99 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Internships\Testing\Integrity;
+
+use Internships\Factories\DataFactory;
+use Internships\Factories\FacultyDataFactory;
+use Internships\FileSystem\DirectoryManager;
+use Internships\FileSystem\FileManager;
+use Internships\Services\CsvReader;
+use Internships\Services\DataSanitizer;
+use Internships\Services\DataValidator;
+use Internships\Services\UniquePathGuard;
+use PHPUnit\Framework\TestCase;
+
+class FromCsvTestCase extends TestCase
+{
+    protected FileManager $fileManager;
+    protected CsvReader $csvReader;
+    protected DirectoryManager $directoryManager;
+    protected FacultyDataFactory $facultyFactory;
+    protected DataValidator $validator;
+
+    protected function setUp(): void
+    {
+        $this->directoryManager = new DirectoryManager(
+            rootDirectoryPath: __DIR__ . "/../../",
+            relativeApiPath: "/public/api/",
+            relativeResourcePath: "/resources/"
+        );
+        $this->fileManager = new FileManager($this->directoryManager, new UniquePathGuard());
+        $this->validator = new DataValidator(new DataSanitizer());
+    }
+
+    protected function checkResourcePath(string $relativePath, string $fileName, bool $checkWrite = false): void
+    {
+        $path = $this->directoryManager->getResourceFilePath($relativePath, $fileName);
+        if ($fileName === "") {
+            $this->assertDirectoryExists($path);
+        } else {
+            $this->assertFileExists($path);
+            if ($checkWrite) {
+                $this->assertFileIsWritable($path);
+            }
+        }
+    }
+
+    protected function retrieveWithFactory(
+        string $relativeDirectory,
+        DataFactory $dataFactory,
+        bool $allowEmptyResult = false
+    ): array {
+        $factoryName = $dataFactory::class;
+        $this->assertTrue(
+            is_subclass_of($dataFactory, DataFactory::class),
+            message: "Class {$factoryName} is not a data factory."
+        );
+
+        $path = $relativeDirectory . $dataFactory->getBaseSourcePath();
+        $filename = $dataFactory->getSourceFileName();
+        $this->checkResourcePath($path, $filename);
+
+        $fields = $dataFactory->getFields();
+        $csvReader = new CsvReader($this->directoryManager, $this->fileManager);
+        $csvData = $csvReader->getCSVData($path, $filename, $fields);
+
+        $this->assertIsArray($csvData[0], message: "Failed creation of array from CSV file.");
+        $this->assertCount(
+            count($fields),
+            array_keys($csvData[0]),
+            message: "{$path}{$filename} has different number of fields than {$factoryName}"
+        );
+
+        $data = $dataFactory->buildFromData($csvData);
+        $this->assertIsArray($data);
+        if (!$allowEmptyResult) {
+            $this->assertNotEmpty(
+                $data,
+                message: "${factoryName} shouldn't return empty array for built data."
+            );
+        }
+        $expectedId = 0;
+        foreach ($data as $modelObject) {
+            $this->assertInstanceOf($dataFactory->getModelClassToBuild(), $modelObject);
+            $this->assertTrue(
+                method_exists($modelObject::class, "getId"),
+                message: "Model generated by {$factoryName} has no required getId method."
+            );
+            $this->assertSame(
+                $expectedId++,
+                $modelObject->getId(),
+                message: "Factory failed assignment of identifiers"
+            );
+        }
+
+        return $data;
+    }
+}
